chore: decomission facetGroup from datasetsSearch
