--- src/main/java/io/github/genomicdatainfrastructure/discovery/datasets/application/ports/FacetsBuilder.java
@@ -1,13 +1,0 @@
-// SPDX-FileCopyrightText: 2024 PNED G.I.E.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package io.github.genomicdatainfrastructure.discovery.datasets.application.ports;
-
-import io.github.genomicdatainfrastructure.discovery.model.DatasetSearchQuery;
-import io.github.genomicdatainfrastructure.discovery.model.FacetGroup;
-
-public interface FacetsBuilder {
-
-    FacetGroup build(DatasetSearchQuery query, String accessToken);
-}
