+++ src/test/java/io/github/genomicdatainfrastructure/discovery/api/RetrieveFacetsIT.java
@@ -1,0 +1,11 @@
+// SPDX-FileCopyrightText: 2024 PNED G.I.E.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package io.github.genomicdatainfrastructure.discovery.api;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class RetrieveFacetsIT extends RetrieveFacetsTest {
+}
